[Radix Cache] Isolate TreeCore test inspection interface
